pyproject.toml
[build-system]
requires = ["setuptools>=80,<81"]
build-backend = "setuptools.build_meta"

[project]
name = "django-amqp"
dynamic = ["version"]
authors = [
  {name = "Karl Hobley", email = "karl@kaed.uk"},
]
requires-python = ">= 3.12"
dependencies = [
  "azure-servicebus>=7.14.3",
  "Django (>=6.0,<7)",
  "pydantic>=2.12.5",
]
description = "AMQP support for django. Supporting RabbitMQ and Azure ServiceBus."
license = "BSD-3-Clause"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.setuptools.dynamic]
version = {attr = "django_amqp.__version__"}

[tool.setuptools.packages.find]
include = ["django_amqp*"]
